Addressing Ana's comments:
please replace the k_nearest_simd_deprecated.c with the
k_nearest_simd_book.c
- please check the Makefile: it is not left with the acheron settings
enabled. I think we can remove those and leave only the first two lines?
It seems to work out of the box on DAS, so this should be alright ... I
have added mine - changed the comments and added a clean option for the
simd in the book version.
- I have added 2 lines in the students Makefile to accommodate for the
book version of SIMD
- In driver.py - maybe you can replace ON_DAS4 with ON_DAS5 (and das4
        with das5). I also noticed that the format of the printing in
the table
does not have a fixed number of decimals (I managed to get numbers like
        4.99000000000004) - this is worth fixing if it is not too much
trouble.
- In all 4 reference codes, please change the text before
opt_classify_MD with "NO NEED to modify this function!"
- please rename the folder to "optimizationlab" - if possible.

-Fixed column's alignment in table printed by driver.py